move shows to store module
